@@ -13,43 +13,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.2",
-    "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^4.3.2",
-    "@angular/compiler": "^4.3.2",
-    "@angular/compiler-cli": "^4.3.2",
-    "@angular/core": "^4.3.2",
-    "@angular/forms": "^4.3.2",
-    "@angular/http": "^4.3.2",
-    "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^4.3.2",
-    "@angular/platform-browser-dynamic": "^4.3.2",
-    "@angular/platform-server": "^4.3.2",
-    "@angular/router": "^4.3.2",
+    "@angular-devkit/build-angular": "^0.13.0",
+    "@angular-devkit/core": "^7.3.0",
+    "@angular/animations": "7.2.3",
+    "@angular/cdk": "^7.3.1",
+    "@angular/common": "7.2.3",
+    "@angular/compiler": "7.2.3",
+    "@angular/compiler-cli": "7.2.3",
+    "@angular/core": "7.2.3",
+    "@angular/forms": "7.2.3",
+    "@angular/http": "7.2.3",
+    "@angular/material": "^7.3.1",
+    "@angular/platform-browser": "7.2.3",
+    "@angular/platform-browser-dynamic": "7.2.3",
+    "@angular/platform-server": "7.2.3",
+    "@angular/router": "7.2.3",
     "core-js": "^2.4.1",
     "hammerjs": "^2.0.8",
-    "json-server": "^0.11.2",
-    "rxjs": "^5.4.2",
+    "json-server": "^0.14.2",
+    "rxjs": "^6.4.0",
     "ts-helpers": "^1.1.2",
-    "typescript": "^2.3.1",
-    "zone.js": "^0.8.16"
+    "typescript": "3.2.4",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "^1.2.6",
+    "@angular/cli": "^7.3.0",
     "@types/jasmine": "^2.5.53",
     "codelyzer": "^3.1.2",
-    "concurrently": "^3.5.0",
+    "concurrently": "^3.6.1",
     "jasmine-core": "^2.7.0",
     "jasmine-spec-reporter": "^4.1.1",
     "karma": "1.7.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "^1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-jasmine": "^1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
-    "protractor": "^5.1.2",
+    "protractor": "^5.4.2",
     "ts-node": "^3.3.0",
     "tslint": "^5.5.0",
-    "typescript": "^2.4.2"
+    "typescript": "3.2.4"
   }
 }
