@@ -1,23 +1,19 @@
-import { Http, Headers } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Item } from './item.model';
-import 'rxjs/add/operator/map';
 
 const BASE_URL = 'http://localhost:3000/items/';
-const HEADER = { headers: new Headers({ 'Content-Type': 'application/json' }) };
 
 @Injectable()
 export class ItemsService {
-  constructor(private http: Http) {}
+  constructor(private http: HttpClient) {}
 
   loadItems() {
-    return this.http.get(BASE_URL)
-      .map(res => res.json());
+    return this.http.get(BASE_URL);
   }
 
   loadItem(id) {
-    return this.http.get(`${BASE_URL}${id}`)
-      .map(res => res.json());
+    return this.http.get(`${BASE_URL}${id}`);
   }
 
   saveItem(item: Item) {
@@ -25,17 +21,14 @@ export class ItemsService {
   }
 
   createItem(item: Item) {
-    return this.http.post(`${BASE_URL}`, JSON.stringify(item), HEADER)
-      .map(res => res.json());
+    return this.http.post(`${BASE_URL}`, item);
   }
 
   updateItem(item: Item) {
-    return this.http.patch(`${BASE_URL}${item.id}`, JSON.stringify(item), HEADER)
-      .map(res => res.json());
+    return this.http.patch(`${BASE_URL}${item.id}`, item);
   }
 
   deleteItem(item: Item) {
-    return this.http.delete(`${BASE_URL}${item.id}`)
-      .map(res => res.json());
+    return this.http.delete(`${BASE_URL}${item.id}`);
   }
 }
